Module Process.Fork_action

Setup actions to perform in the child process.

type t = Eio_unix.Private.Fork_action.t

val execve : string -> argv:string array -> env:string array -> t

See execve(2). This replaces the current executable, so it only makes sense as the last action to be performed.

val chdir : string -> t

chdir path changes directory to path.

val fchdir : FD.t -> t

fchdir dir changes directory to dir.